Got the data coming from the poppler stub correctly showing in gui.
Changed some atribute names for more consistency.
Changed data to time and the type is now gchar instead of GTime.

And the segfault magically dissapeared :D.
It shall come to annoy us in the future...

Adding "g_object_unref (job)" at the end of job_finished_callback
gets some warnings while running so not adding that for now.