Show info message regarding the presence of signatures.
So, made it simpler.
We could have shown two different information messages:
 * one saying that everything was good with the signatures;
 * other one saying that there were problems.

This would require that we excuted the verification part
of the signatures just to show this info and I can't see
a simple way to decouple that logic from the sidebar for now.

So, instead of all that complicated stuff, just show a
message saying there are signatures and offer the option
to see the sidebar with all the information.