Changed the logic on the gui about the status a little.
Now the icons map in a more logic way to the status of the document
* document completly ok;                     (o)
* document ok but with problems              (!)
* document shouldn't be trusted, is damaged  (x)

Changed the info that the backend sends to our dummys just to test all
this new logic. Seems to work.

Also added/changed the way the backend will get the info considering
the future changes in the poppler-glib API related to the signatures.